requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.8.2
blis==0.7.11
Brotli==1.1.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.8
Cython==0.29.37
defusedxml==0.7.1
dnspython==2.6.1
duckdb==1.0.0
eli5==0.13.0
email_validator==2.2.0
fastai==2.7.15
fastapi==0.111.0
fastapi-cli==0.0.4
fastcore==1.5.54
fastdownload==0.0.7
fastjsonschema==2.17.1
fastprogress==1.0.3
featuretools==1.31.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gevent==24.2.1
geventhttpclient==2.3.1
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
hdbscan==0.8.37
holidays==0.52
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.1
imbalanced-learn==0.12.2
imblearn==0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.4.2
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.2.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
keras==3.4.1
kiwisolver==1.4.4
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
lazypredict==0.2.12
libclang==18.1.1
lightgbm==4.3.0
lightning==2.3.3
lightning-utilities==0.11.5
lime==0.2.0.1
llvmlite==0.42.0
locust==2.29.1
Mako==1.3.3
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.1
mkl==2021.4.0
mkl-service==2.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
namex==0.0.8
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.1
nbformat==5.9.1
networkx==3.3
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.59.1
numexpr==2.8.4
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
optuna==3.6.1
orjson==3.10.6
overrides==7.3.1
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pillow==10.0.0
platformdirs==3.0.0
ply==3.11
polars==1.0.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==4.25.4
psutil==6.0.0
pyarrow==16.1.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.16.1
pyparsing==3.0.9
PyQt5==5.15.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-mnist==0.7
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2023.3
pywin32==305.1
pywinpty==2.0.11
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
regex==2024.4.28
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.8.11
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.2
shap==0.45.1
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smart-open==7.0.4
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
statsmodels==0.14.2
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.5.3